Fix alignment of numbers in post list view
Currently, going from single digit numbers to double digit
numbers pushes the post title to the right, making it look
uneven. This sets a minimum width and uses a monospaced
font in order to stop that, although this is too rigid to
work for three digit numbers. Seems to be rare that a user
ever gets to >99 posts loaded though.